Method by which the next parameter is selected is based on a "mixin" class
In the past, the assumption has been that we would always use the method of
constant limits, so that was built-in to the paradigm.  However, I am moving
away from this model towards one in which we can provide different "methods"
via mixin classes.  Maximum likelihood is one method.  We will certianly have a
constant limits method as well; however, that won't be implemented quite yet.

Towards this goal, I am revising the paradigms to remove any reference to the
parameters that will be tested.  We would only enter a sequence if we are using
the method of constant limits, so the paradigm is not the place to contain this
information.  This information should be contained in the method of constant
limits mixin.