Physiology circuit sampling frequency changed to 12 kHz
Current physiology data files are ~5 GB for 30 minutes of data.  By cutting
sampling frequency in half, we cut file size in half.  Since we are filtering
the waveform between 300 and 5000 Hz, it is a bit silly to sample at higher
frequencies.

Note that incoming signal from Medusa is now lowpass filtered with a maximum
lowpass cutoff frequency of 5000 Hz.  This is hard-coded to avoid aliasing
effects.  The user may reduce the filter bandwidth further if they desire.