Some cleanup of the documentation
Neurobehavior documentation is currently lacking.  Much of the code that was
actually documented has been moved to their own modules (e.g. TDTPy and New Era)
so I have removed the documentation that pertains to this code.  As for the rest
of the documentation, some of this is better described in the writeup on Google
docs because it is lab-specific rather than specific to Neurobehavior.  I have
removed this documentation as well.

As for the rest, I will eventually get around to updating it; however, the focus
will likely be on working examples that people can play with.