Updated to utitilize the RPC version TDTPy
A recent change to TDTPy allows hardware I/O to be run in a separate process.
This is a transparent change for the most part.  If a host/port is not specified
then the hardware I/O is run in the same process as before.  If a host/port is
specified, then the RPC client version of TDTPy will be used.