Bugfixes to channel plots under cns.chaco_exts
Various changes in the cns.chaco_exts to get the off-line physiology analysis
working broke some of the on-line analysis code (mainly the code that detects
changes to the underlying data sources and updates the plot accordingly).  This
changeset contains the requisite bugfixes.

One key change was updating all plot components to refer to a "source" attribute
than a channel, series or epoch attribute.  However, the exact class the source
attribute requires depends on the plot type (e.g. the source for an EpochPlot
requires a cns.channel.Epoch class and cannot support a MultiChannel).

Additional changes:
* Modified ChannelRangeTool so you can disable zooming or dragging
* Updated positive experiment so you can control-scroll with the mouse wheel to
  zoom in/out on the experiment overview plot.