Additional progress towards refactoring experiment paradigms (NOT COMPLETE)
This is not a funcitonal refactor, but is meant as a "safety commit".  Pending
items include

* Completing the refactor to include determining the appropriate attenuation and
  scaling factor for the waveform
* Moving the speaker/equalization and fixed attenuation logic out of the
  abstract controller classes.  This is an overly complex feature and is not
  necessarily shared among all paradigms.

We are converging on a very clear idea of the best way to set up experiment
paradigms so that they are robust, easy to understand and easy to write.  This
includes unwinding all dependencies on Neurogen and moving the computation of
the waveforms to the relevant paradigms/paradigm_name.py files (rather than
hiding them in the Neurogen block logic).

All utility functions for generating signals are currently in cns/signal.py.

Note that we still have PositiveAMNoise{Controller|Paradigm}Mixin and
PositiveDT{Controller|Paradigm}Mixin classes, but these have been moved to the
paradigms folder and prepended with an underscore to indicate the fact that
these are meant to be shared by the training and actual paradigms, respectively.

This commit also includes removal of unused imports.