Modified by Antje to remove initial "click" generated when circuit starts
The click was coming from the masker playing as soon as the circuit was started.
Since the masker is meant to be continuous, there's no onset ramp.  By setting
masker_scale tag in the RPvds circuit to 0 by default, we can eliminate this
transient (the Python code will slowly ramp up the masker_scale tag from the
default value of 0 to the final scaling factor required over the course of ~1-2
msec).