Experimental version of loader.py that stores experiment data in separate files
The current cohort.hd5 file approach stores all data in a single monolithic
file.  If an experiment crashes, there's potential for corrupting the entire
file (i.e. the entire cohort dataset).  Although the probability of corruption
(aside from the hard drive filling up) is extremely low, we created a new
version that stores each experiment in a separate HDF5 file and links back
to the main HDF5 file.  However, I can't get the ExternalLink to work robustly
(it appears to break the link if you move the parent/child data to a new
folder).