Bugfix to ProcessedMultiChannel problem reported by GvT
During class initialization, the default value for the number of channels was
8. Before the default value gets updated to the correct value of 16, the
differential matrix was getting computed. This generally is fine; however, if
you happen to have exactly 7 bad channels, this was creating a
ZeroDivisionError.

By setting the default value to 0, this avoids the ZeroDivisionError and makes
it clear that the class has not finished initializing.  The correct differential
will be re-computed when the correct value for channel number is set.